# specs/001-ai-dashboard-backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Joyfull UI Hub Backend API
  description: |
    Backend API for multi-tool AI dashboard with authentication, user management,
    tool access, usage tracking, and subscription management.
  version: 1.0.0
  contact:
    name: Joyfull UI Hub Team
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /tools:
    get:
      tags: [tools]
      summary: List available AI tools
      operationId: listTools
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: List of tools
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ToolSummary'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tools/{tool_id}:
    get:
      tags: [tools]
      summary: Get tool details
      operationId: getTool
      security:
        - bearerAuth: []
      parameters:
        - name: tool_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tool details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ToolDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /tools/{tool_id}/invoke:
    post:
      tags: [tools]
      summary: Invoke an AI tool
      operationId: invokeTool
      security:
        - bearerAuth: []
      parameters:
        - name: tool_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input]
              properties:
                input:
                  type: object
                  description: Tool-specific input payload
      responses:
        '200':
          description: Tool invocation result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ToolInvocationResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Subscription required or token limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          description: Tool service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  # ============================================================================
  # Usage Endpoints
  # ============================================================================
  /usage:
    get:
      tags: [usage]
      summary: Get user's usage history
      operationId: getUsageHistory
      security:
        - bearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: tool_id
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Usage history
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/UsageLogEntry'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /usage/summary:
    get:
      tags: [usage]
      summary: Get aggregated usage summary
      operationId: getUsageSummary
      security:
        - bearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Usage summary
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UsageSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================================
  # Subscription Endpoints
  # ============================================================================
  /subscriptions/plans:
    get:
      tags: [subscriptions]
      summary: List available subscription plans
      operationId: listPlans
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Available plans
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/SubscriptionPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /subscriptions/current:
    get:
      tags: [subscriptions]
      summary: Get current subscription
      operationId: getCurrentSubscription
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current subscription
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubscriptionDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /subscriptions/subscribe:
    post:
      tags: [subscriptions]
      summary: Subscribe to a plan
      operationId: subscribe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan]
              properties:
                plan:
                  type: string
                  enum: [free, pro, enterprise]
      responses:
        '200':
          description: Subscription updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubscriptionDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /subscriptions/cancel:
    post:
      tags: [subscriptions]
      summary: Cancel current subscription
      operationId: cancelSubscription
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subscription cancelled
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubscriptionDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================================
  # Admin Endpoints
  # ============================================================================
...
  /admin/tools:
    post:
      tags: [admin]
      summary: Create a new tool (admin only)
      operationId: adminCreateTool
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolCreate'
      responses:
        '201':
          description: Tool created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ToolDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/tools/{tool_id}:
    patch:
      tags: [admin]
      summary: Update a tool (admin only)
      operationId: adminUpdateTool
      security:
        - bearerAuth: []
      parameters:
        - name: tool_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolUpdate'
      responses:
        '200':
          description: Tool updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ToolDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [admin]
      summary: Delete a tool (admin only)
      operationId: adminDeleteTool
      security:
        - bearerAuth: []
      parameters:
        - name: tool_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Tool deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    ApiResponse:
      type: object
      properties:
        data:
          nullable: true
        error:
          nullable: true
        meta:
          type: object
          nullable: true
    ApiError:
      type: object
      properties:
        data:
          nullable: true
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
        meta:
          type: object
          nullable: true
    PaginationMeta:
      type: object
      properties:
        pagination:
          type: object
          properties:
            offset:
              type: integer
            limit:
              type: integer
            total:
              type: integer

    # Auth schemas
...
    ToolSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    ToolDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    ToolCreate:
      type: object
      required: [name, description, api_endpoint]
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
        api_endpoint:
          type: string
          format: uri
    ToolUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
        api_endpoint:
          type: string
          format: uri
        is_active:
          type: boolean
    ToolInvocationResult:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        output:
          type: object
        tokens_used:
          type: integer
        duration_ms:
          type: integer

    # Usage schemas
    UsageLogEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tool_id:
          type: string
          format: uuid
        tool_name:
          type: string
        timestamp:
          type: string
          format: date-time
        tokens_used:
          type: integer
        response_status:
          type: string
          enum: [success, error, timeout]
    UsageSummary:
      type: object
      properties:
        total_tokens:
          type: integer
        total_invocations:
          type: integer
        by_tool:
          type: array
          items:
            type: object
            properties:
              tool_id:
                type: string
                format: uuid
              tool_name:
                type: string
              tokens_used:
                type: integer
              invocation_count:
                type: integer

    # Subscription schemas
    SubscriptionPlan:
      type: object
      properties:
        name:
          type: string
          enum: [free, pro, enterprise]
        token_limit:
          type: integer
        features:
          type: array
          items:
            type: string
    SubscriptionDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        plan:
          type: string
          enum: [free, pro, enterprise]
        status:
          type: string
          enum: [active, cancelled, expired]
        token_limit:
          type: integer
        tokens_used:
          type: integer
        tokens_remaining:
          type: integer
        period_start:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true

    # Admin schemas
...